+++ tests/DownloadedFile/DownloadedFileTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phptg\BotApi\Tests;
+namespace Phptg\BotApi\Tests\DownloadedFile;
 
 use PHPUnit\Framework\TestCase;
 use Phptg\BotApi\DownloadedFile;
@@ -32,25 +32,6 @@ final class DownloadedFileTest extends TestCase
         $file = new DownloadedFile($stream);
 
         assertSame('hello-content', $file->getBody());
-    }
-
-    public function testSaveTo(): void
-    {
-        $stream = fopen('php://temp', 'r+b');
-        fwrite($stream, 'hello-content');
-        rewind($stream);
-
-        $file = new DownloadedFile($stream);
-
-        $path = sys_get_temp_dir() . '/downloaded-file-test-' . uniqid(more_entropy: true) . '.txt';
-        try {
-            $file->saveTo($path);
-            assertSame('hello-content', file_get_contents($path));
-        } finally {
-            if (file_exists($path)) {
-                unlink($path);
-            }
-        }
     }
 
     public function testSaveToError(): void
